# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ginutil.Result:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
  repo.ChartVersion:
    properties:
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations are additional mappings uninterpreted by Tiller,
          made available for inspection by other applications.
        type: object
      apiVersion:
        description: The API Version of this chart.
        type: string
      appVersion:
        description: The version of the application enclosed inside of this chart.
        type: string
      condition:
        description: The condition to check to enable chart
        type: string
      created:
        type: string
      deprecated:
        description: Whether or not this chart is deprecated
        type: boolean
      description:
        description: A one-sentence description of the chart
        type: string
      digest:
        type: string
      engine:
        description: The name of the template engine to use. Defaults to 'gotpl'.
        type: string
      home:
        description: The URL to a relevant project page, git repo, or contact person
        type: string
      icon:
        description: The URL to an icon file.
        type: string
      keywords:
        description: A list of string keywords
        items:
          type: string
        type: array
      kubeVersion:
        description: KubeVersion is a SemVer constraint specifying the version of
          Kubernetes required.
        type: string
      maintainers:
        description: A list of name and URL/email address combinations for the maintainer(s)
        items:
          $ref: '#/definitions/chart.Maintainer'
        type: array
      name:
        description: The name of the chart
        type: string
      removed:
        type: boolean
      sources:
        description: Source is the URL to the source code of this chart
        items:
          type: string
        type: array
      tags:
        description: The tags to check to enable chart
        type: string
      tillerVersion:
        description: |-
          TillerVersion is a SemVer constraints on what version of Tiller is required.
          See SemVer ranges here: https://github.com/Masterminds/semver#basic-comparisons
        type: string
      urls:
        items:
          type: string
        type: array
      version:
        description: A SemVer 2 conformant version string of the chart
        type: string
    type: object
...
  v1.AppStore:
    properties:
      branch:
        description: The branch of app store, which category is git repo.
        type: string
      eid:
        description: The enterprise id.
        type: string
      name:
        description: The name of app store.
        type: string
      password:
        description: The password of the private app store
        type: string
      url:
        description: The url of app store.
        type: string
      username:
        description: The username of the private app store
        type: string
    type: object
  v1.AppTemplate:
    properties:
      name:
        description: The name of app template.
        type: string
      versions:
        description: A list of app template versions.
        items:
          $ref: '#/definitions/repo.ChartVersion'
        type: array
    type: object
...
  v1.CreateAppStoreReq:
    properties:
      branch:
        description: The branch of app store, which category is git repo.
        type: string
      name:
        description: The name of app store.
        type: string
      password:
        description: The password of the private app store
        type: string
      url:
        description: The url of app store.
        type: string
      username:
        description: The username of the private app store
        type: string
    required:
    - name
    - url
    type: object
...
  v1.EventSeries:
    properties:
      count:
        description: Number of occurrences in this series up to the last heartbeat
          time
        type: integer
      lastObservedTime:
        $ref: '#/definitions/v1.MicroTime'
        description: Time of the last occurrence observed
    type: object
...
  v1.FieldsV1:
    type: object
...
  v1.ManagedFieldsEntry:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the version of this resource that this field set
          applies to. The format is "group/version" just like the top-level
          APIVersion field. It is necessary to track the version of a field
          set because it cannot be automatically converted.
        type: string
      fieldsType:
        description: |-
          FieldsType is the discriminator for the different fields format and version.
          There is currently only one possible value: "FieldsV1"
        type: string
      fieldsV1:
        $ref: '#/definitions/v1.FieldsV1'
        description: |-
          FieldsV1 holds the first JSON version format as described in the "FieldsV1" type.
          +optional
      manager:
        description: Manager is an identifier of the workflow managing these fields.
        type: string
      operation:
        description: |-
          Operation is the type of operation which lead to this ManagedFieldsEntry being created.
          The only valid values for this field are 'Apply' and 'Update'.
        type: string
      time:
        description: |-
          Time is timestamp of when these fields were set. It should always be empty if Operation is 'Apply'
          +optional
        type: string
    type: object
  v1.MicroTime:
    properties:
      time.Time:
        type: string
    type: object
...
  v1.ObjectReference:
    properties:
      apiVersion:
        description: |-
          API version of the referent.
          +optional
        type: string
      fieldPath:
        description: |-
          If referring to a piece of an object instead of an entire object, this string
          should contain a valid JSON/Go field access statement, such as desiredState.manifest.containers[2].
          For example, if the object reference is to a container within a pod, this would take on a value like:
          "spec.containers{name}" (where "name" refers to the name of the container that triggered
          the event) or if no container name is specified "spec.containers[2]" (container with
          index 2 in this pod). This syntax is chosen only to have some well-defined way of
          referencing a part of an object.
          TODO: this design is not final and this field is subject to change in the future.
          +optional
        type: string
      kind:
        description: |-
          Kind of the referent.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      name:
        description: |-
          Name of the referent.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names
          +optional
        type: string
      namespace:
        description: |-
          Namespace of the referent.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces/
          +optional
        type: string
      resourceVersion:
        description: |-
          Specific resourceVersion to which this reference is made, if any.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency
          +optional
        type: string
      uid:
        description: |-
          UID of the referent.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#uids
          +optional
        type: string
    type: object
...
  v1.PruneUpdateRKEConfigReq:
    properties:
      encodedRKEConfig:
        type: string
      nodes:
        items:
          $ref: '#/definitions/v1alpha1.ConfigNode'
        type: array
    type: object
  v1.PruneUpdateRKEConfigResp:
    properties:
      encodeRKEConfig:
        type: string
      nodes:
        items:
          $ref: '#/definitions/v1alpha1.ConfigNode'
        type: array
    type: object
...
  v1.WutongComponent:
    properties:
      app:
        type: string
      pods:
        items:
          $ref: '#/definitions/v1.Pod'
        type: array
    type: object
  v1.WutongComponentEvent:
    properties:
      action:
        description: |-
          What action was taken/failed regarding to the Regarding object.
          +optional
        type: string
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations is an unstructured key value map stored with a resource that may be
          set by external tools to store and retrieve arbitrary metadata. They are not
          queryable and should be preserved when modifying objects.
          More info: http://kubernetes.io/docs/user-guide/annotations
          +optional
        type: object
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      clusterName:
        description: |-
          The name of the cluster which the object belongs to.
          This is used to distinguish resources with same name and namespace in different clusters.
          This field is not set anywhere right now and apiserver is going to ignore it if set in create or update request.
          +optional
        type: string
      count:
        description: |-
          The number of times this event has occurred.
          +optional
        type: integer
      creationTimestamp:
        description: |-
          CreationTimestamp is a timestamp representing the server time when this object was
          created. It is not guaranteed to be set in happens-before order across separate operations.
          Clients may not set this value. It is represented in RFC3339 form and is in UTC.

          Populated by the system.
          Read-only.
          Null for lists.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      deletionGracePeriodSeconds:
        description: |-
          Number of seconds allowed for this object to gracefully terminate before
          it will be removed from the system. Only set when deletionTimestamp is also set.
          May only be shortened.
          Read-only.
          +optional
        type: integer
      deletionTimestamp:
        description: |-
          DeletionTimestamp is RFC 3339 date and time at which this resource will be deleted. This
          field is set by the server when a graceful deletion is requested by the user, and is not
          directly settable by a client. The resource is expected to be deleted (no longer visible
          from resource lists, and not reachable by name) after the time in this field, once the
          finalizers list is empty. As long as the finalizers list contains items, deletion is blocked.
          Once the deletionTimestamp is set, this value may not be unset or be set further into the
          future, although it may be shortened or the resource may be deleted prior to this time.
          For example, a user may request that a pod is deleted in 30 seconds. The Kubelet will react
          by sending a graceful termination signal to the containers in the pod. After that 30 seconds,
          the Kubelet will send a hard termination signal (SIGKILL) to the container and after cleanup,
          remove the pod from the API. In the presence of network partitions, this object may still
          exist after this timestamp, until an administrator or automated process can determine the
          resource is fully terminated.
          If not set, graceful deletion of the object has not been requested.

          Populated by the system when a graceful deletion is requested.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      eventTime:
        $ref: '#/definitions/v1.MicroTime'
        description: |-
          Time when this Event was first observed.
          +optional
      finalizers:
        description: |-
          Must be empty before the object is deleted from the registry. Each entry
          is an identifier for the responsible component that will remove the entry
          from the list. If the deletionTimestamp of the object is non-nil, entries
          in this list can only be removed.
          Finalizers may be processed and removed in any order.  Order is NOT enforced
          because it introduces significant risk of stuck finalizers.
          finalizers is a shared field, any actor with permission can reorder it.
          If the finalizer list is processed in order, then this can lead to a situation
          in which the component responsible for the first finalizer in the list is
          waiting for a signal (field value, external system, or other) produced by a
          component responsible for a finalizer later in the list, resulting in a deadlock.
          Without enforced ordering finalizers are free to order amongst themselves and
          are not vulnerable to ordering changes in the list.
          +optional
          +patchStrategy=merge
        items:
          type: string
        type: array
      firstTimestamp:
        description: |-
          The time at which the event was first recorded. (Time of server receipt is in TypeMeta.)
          +optional
        type: string
      generateName:
        description: |-
          GenerateName is an optional prefix, used by the server, to generate a unique
          name ONLY IF the Name field has not been provided.
          If this field is used, the name returned to the client will be different
          than the name passed. This value will also be combined with a unique suffix.
          The provided value has the same validation rules as the Name field,
          and may be truncated by the length of the suffix required to make the value
          unique on the server.

          If this field is specified and the generated name exists, the server will
          NOT return a 409 - instead, it will either return 201 Created or 500 with Reason
          ServerTimeout indicating a unique name could not be found in the time allotted, and the client
          should retry (optionally after the time indicated in the Retry-After header).

          Applied only if Name is not specified.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#idempotency
          +optional
        type: string
      generation:
        description: |-
          A sequence number representing a specific generation of the desired state.
          Populated by the system. Read-only.
          +optional
        type: integer
      involvedObject:
        $ref: '#/definitions/v1.ObjectReference'
        description: The object that this event is about.
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      labels:
        additionalProperties:
          type: string
        description: |-
          Map of string keys and values that can be used to organize and categorize
          (scope and select) objects. May match selectors of replication controllers
          and services.
          More info: http://kubernetes.io/docs/user-guide/labels
          +optional
        type: object
      lastTimestamp:
        description: |-
          The time at which the most recent occurrence of this event was recorded.
          +optional
        type: string
      managedFields:
        description: |-
          ManagedFields maps workflow-id and version to the set of fields
          that are managed by that workflow. This is mostly for internal
          housekeeping, and users typically shouldn't need to set or
          understand this field. A workflow can be the user's name, a
          controller's name, or the name of a specific apply path like
          "ci-cd". The set of fields is always in the version that the
          workflow used when modifying the object.

          +optional
        items:
          $ref: '#/definitions/v1.ManagedFieldsEntry'
        type: array
      message:
        description: |-
          A human-readable description of the status of this operation.
          TODO: decide on maximum length.
          +optional
        type: string
      name:
        description: |-
          Name must be unique within a namespace. Is required when creating resources, although
          some resources may allow a client to request the generation of an appropriate name
          automatically. Name is primarily intended for creation idempotence and configuration
          definition.
          Cannot be updated.
          More info: http://kubernetes.io/docs/user-guide/identifiers#names
          +optional
        type: string
      namespace:
        description: |-
          Namespace defines the space within which each name must be unique. An empty namespace is
          equivalent to the "default" namespace, but "default" is the canonical representation.
          Not all objects are required to be scoped to a namespace - the value of this field for
          those objects will be empty.

          Must be a DNS_LABEL.
          Cannot be updated.
          More info: http://kubernetes.io/docs/user-guide/namespaces
          +optional
        type: string
      ownerReferences:
        description: |-
          List of objects depended by this object. If ALL objects in the list have
          been deleted, this object will be garbage collected. If this object is managed by a controller,
          then an entry in this list will point to this controller, with the controller field set to true.
          There cannot be more than one managing controller.
          +optional
          +patchMergeKey=uid
          +patchStrategy=merge
        items:
          $ref: '#/definitions/v1.OwnerReference'
        type: array
      reason:
        description: |-
          This should be a short, machine understandable string that gives the reason
          for the transition into the object's current status.
          TODO: provide exact specification for format.
          +optional
        type: string
      related:
        $ref: '#/definitions/v1.ObjectReference'
        description: |-
          Optional secondary object for more complex actions.
          +optional
      reportingComponent:
        description: |-
          Name of the controller that emitted this Event, e.g. `kubernetes.io/kubelet`.
          +optional
        type: string
      reportingInstance:
        description: |-
          ID of the controller instance, e.g. `kubelet-xyzf`.
          +optional
        type: string
      resourceVersion:
        description: |-
          An opaque value that represents the internal version of this object that can
          be used by clients to determine when objects have changed. May be used for optimistic
          concurrency, change detection, and the watch operation on a resource or set of resources.
          Clients must treat these values as opaque and passed unmodified back to the server.
          They may only be valid for a particular resource or set of resources.

          Populated by the system.
          Read-only.
          Value must be treated as opaque by clients and .
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency
          +optional
        type: string
      selfLink:
        description: |-
          SelfLink is a URL representing this object.
          Populated by the system.
          Read-only.

          DEPRECATED
          Kubernetes will stop propagating this field in 1.20 release and the field is planned
          to be removed in 1.21 release.
          +optional
        type: string
      series:
        $ref: '#/definitions/v1.EventSeries'
        description: |-
          Data about the Event series this event represents or nil if it's a singleton Event.
          +optional
      source:
        $ref: '#/definitions/v1.EventSource'
        description: |-
          The component reporting this event. Should be a short machine understandable string.
          +optional
      type:
        description: |-
          Type of this event (Normal, Warning), new types could be added in the future
          +optional
        type: string
      uid:
        description: |-
          UID is the unique in time and space value for this object. It is typically generated by
          the server on successful creation of a resource and is not allowed to change on PUT
          operations.

          Populated by the system.
          Read-only.
          More info: http://kubernetes.io/docs/user-guide/identifiers#uids
          +optional
        type: string
    type: object
...
  v1.UpdateKubernetesReq:
    properties:
      clusterID:
        type: string
      encodedRKEConfig:
        type: string
      etcdNodeNum:
        type: integer
      instanceType:
        type: string
      masterNodeNum:
        type: integer
      provider:
        type: string
      workerNum:
        type: integer
      workerResourceType:
        type: string
    type: object
  v1.UpdateKubernetesTask:
    properties:
      clusterID:
        type: string
      eid:
        type: string
      nodeNumber:
        type: integer
      providerName:
        type: string
      status:
        type: string
      taskID:
        type: string
    type: object
...
  v1alpha1.ConfigNode:
    properties:
      dockerSocketPath:
        type: string
      internalIP:
        type: string
      ip:
        type: string
      roles:
        items:
          type: string
        type: array
      sshPort:
        type: integer
      sshUser:
        type: string
    type: object
...
info:
  contact:
    name: Wutong
    url: https://wutong.com
  description: Cloud Adaptor
  title: Cloud Adaptor API
  version: "1.0"
paths:
  /api/v1/enterprises/:eid/appstores:
    get:
      consumes:
      - application/json
      operationId: listAppStores
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.AppStore'
            type: array
      summary: returns a list of app stores.
      tags:
      - appstores
    post:
      consumes:
      - application/json
      operationId: createAppStore
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: .
        in: body
        name: createAppStoreReq
        required: true
        schema:
          $ref: '#/definitions/v1.CreateAppStoreReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AppStore'
        "400":
          description: 8002, app store unavailable
          schema:
            $ref: '#/definitions/ginutil.Result'
        "409":
          description: 8001, app store name conflict
          schema:
            $ref: '#/definitions/ginutil.Result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: creates a new app store.
      tags:
      - appstores
...
  /api/v1/enterprises/:eid/appstores/:name/apps:
    get:
      consumes:
      - application/json
      operationId: listTemplates
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the name of the app store
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.AppTemplate'
            type: array
        "404":
          description: 8000, app store not found
          schema:
            $ref: '#/definitions/ginutil.Result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: returns a list of app templates.
      tags:
      - appstores
  /api/v1/enterprises/:eid/appstores/:name/apps/:templateName:
    get:
      consumes:
      - application/json
      operationId: getAppTemplate
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the name of the app store
        in: path
        name: name
        required: true
        type: string
      - description: the name of the app template
        in: path
        name: templateName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AppTemplate'
        "404":
          description: 8000, app store not found
          schema:
            $ref: '#/definitions/ginutil.Result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: returns the app template.
      tags:
      - appstores
...
  /api/v1/enterprises/:eid/kclusters/prune-update-rkeconfig:
    post:
      consumes:
      - application/json
      operationId: pruneUpdateRKEConfig
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: .
        in: body
        name: pruneUpdateRKEConfigReq
        required: true
        schema:
          $ref: '#/definitions/v1.PruneUpdateRKEConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PruneUpdateRKEConfigResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: update rke config purely
      tags:
      - cluster
  /api/v1/enterprises/:eid/update-cluster:
    post:
      consumes:
      - application/json
      operationId: updateKubernetesCluster
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: .
        in: body
        name: updateKubernetesReq
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateKubernetesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UpdateKubernetesTask'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: updates kubernetes cluster.
      tags:
      - cluster
  /api/v1/enterprises/:eid/update-cluster/:clusterID:
    get:
      consumes:
      - application/json
      operationId: getUpdateKubernetesTask
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the cluster id
        in: path
        name: clusterID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UpdateKubernetesTask'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: returns the information of the cluster.
      tags:
      - cluster
  /api/v1/enterprises/{eid}/kclusters/{clusterID}/wutong-components:
    get:
      consumes:
      - application/json
      operationId: listWutongComponents
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the identify of cluster
        in: path
        name: clusterID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.WutongComponent'
            type: array
      summary: returns a list of wutong components.
      tags:
      - cluster
  /api/v1/enterprises/{eid}/kclusters/{clusterID}/wutong-components/{podName}/events:
    get:
      consumes:
      - application/json
      operationId: listPodEvents
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the identify of cluster
        in: path
        name: clusterID
        required: true
        type: string
      - description: the name of pod
        in: path
        name: podName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.WutongComponentEvent'
            type: array
      summary: returns a list of wutong component pod events.
      tags:
      - cluster
schemes:
- http,
- https
swagger: "2.0"